% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_EFGLdata.R
\name{aRich}
\alias{aRich}
\title{calculate allelic richness of loci}
\usage{
aRich(x)
}
\arguments{
\item{x}{an EFGLdata object}
}
\value{
a tibble giving the allelic richness of each locus
}
\description{
calculate allelic richness of loci
}
