% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_EFGLdata.R
\name{calcAF}
\alias{calcAF}
\title{calculate allele frequencies within populations
Loci that are all missing genotypes wihtin a pop will be
(silently) excluded form the output.}
\usage{
calcAF(x)
}
\arguments{
\item{x}{an EFGLdata object}
}
\value{
a tibble
}
\description{
calculate allele frequencies within populations
Loci that are all missing genotypes wihtin a pop will be
(silently) excluded form the output.
}
