% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_EFGLdata.R
\name{calcHet}
\alias{calcHet}
\title{calculate expected and observed heterozygosity within populations}
\usage{
calcHet(x)
}
\arguments{
\item{x}{an EFGLdata object}
}
\value{
a tibble
}
\description{
calculate expected and observed heterozygosity within populations
}
