% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{cleanGrandma}
\alias{cleanGrandma}
\title{convenience function to remove loci with all fails or no variation from
gRandma input}
\usage{
cleanGrandma(baseline, mixture = NULL)
}
\arguments{
\item{baseline}{a gRandma baseline input}

\item{mixture}{a gRandma mixture input}
}
\value{
a list with two components, one is the baseline, one is the mixture
}
\description{
convenience function to remove loci with all fails or no variation from
gRandma input
}
