% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFGLdata.R
\name{construct_EFGLdata}
\alias{construct_EFGLdata}
\title{some basic checks on EFGLdata objects}
\usage{
construct_EFGLdata(x)
}
\arguments{
\item{x}{an EFGLdata object}
}
\description{
some basic checks on EFGLdata objects
}
