% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_EFGLdata.R
\name{dumpTable}
\alias{dumpTable}
\title{wrapper for write table with commonly used options - carried over from IDFGEN}
\usage{
dumpTable(x, filename, row.names = FALSE, sep = "\\t")
}
\arguments{
\item{x}{object to write out}

\item{filename}{filename to write out as}

\item{row.names}{passed to write.table}

\item{sep}{passed to write.table}
}
\description{
wrapper for write table with commonly used options - carried over from IDFGEN
}
