% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFGLdata.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{An example input dataset used in the vignette}
\format{
a tibble
}
\usage{
exampleData
}
\description{
An example input dataset used in the vignette
}
\keyword{datasets}
