% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{exportCKMRsimAF}
\alias{exportCKMRsimAF}
\title{return a CKMRsim allele frequency input tibble.}
\usage{
exportCKMRsimAF(x, pops = NULL, loci = NULL)
}
\arguments{
\item{x}{an EFGLdata object}

\item{pops}{a vector of pops to include. If not specified,
all pops are used.}

\item{loci}{a vector of loci to include. If not specified,
all loci are used.}
}
\value{
a tibble
}
\description{
No chromosome and position information is used.
All populations specified are combined into one allele frequency output.
Any loci with only missing genotypes are omitted from the output. The output
should then be run through the CKMRsim function \code{reindex_markers()}.
}
