% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{exportCKMRsimLG}
\alias{exportCKMRsimLG}
\title{Return a long genotype tibble for input to CKMRsim for relationship inference.}
\usage{
exportCKMRsimLG(x, pops = NULL, loci = NULL)
}
\arguments{
\item{x}{an EFGLdata object}

\item{pops}{a vector of pops to include. If not specified,
all pops are used.}

\item{loci}{a vector of loci to include. If not specified,
all loci are used.}
}
\value{
a tibble
}
\description{
Return a long genotype tibble for input to CKMRsim for relationship inference.
}
