% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{exportColony}
\alias{exportColony}
\title{write a Colony input file.}
\usage{
exportColony(
  x,
  filename = "Colony2.dat",
  pops = NULL,
  loci = NULL,
  candMales = NULL,
  candFemales = NULL,
  candMaleProb = NULL,
  candFemaleProb = NULL,
  projName = "efglmhToColony",
  colOutFileName = "colonyResults",
  alleleDropoutRate = 0.01,
  otherGenotypingErrorRate = 0.01
)
}
\arguments{
\item{x}{an EFGLdata object}

\item{filename}{the name of the file to write. If NULL, the colony input is returned
as a character vector with each line a separate entry in the vector.}

\item{pops}{a vector of pops to include. If not specified,
all pops are used.}

\item{loci}{a vector of loci to include. If not specified,
all loci are used.}

\item{candMales}{candidate male names}

\item{candFemales}{candidate female names}

\item{candMaleProb}{probability father is included in candidate males. if NULL, either 0 or 0.5 depending on whether candMales is NULL.}

\item{candFemaleProb}{probability father is included in candidate Females. if NULL, either
0 or 0.5 depending on whether candFemales is NULL.}

\item{projName}{project name in colony input}

\item{colOutFileName}{output file name to direct colony to use}

\item{alleleDropoutRate}{allelic dropout rate to use for all loci}
}
\value{
nothing, just writes a file
}
\description{
write a Colony input file.
}
