% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{exportGenAlEx}
\alias{exportGenAlEx}
\title{write a GenAlEx input file}
\usage{
exportGenAlEx(
  x,
  filename,
  pops = NULL,
  loci = NULL,
  title = "",
  useNames = TRUE
)
}
\arguments{
\item{x}{an EFGLdata object}

\item{filename}{the name of the file to write}

\item{pops}{a vector of pops to include. If not specified,
all pops are used.}

\item{loci}{a vector of loci to include. If not specified,
all loci are used.}

\item{title}{a string to use as the "title" row}

\item{useNames}{TRUE to use samples names, FALSE to replace with unique numerical identifiers}
}
\value{
nothing, just writes a file
}
\description{
write a GenAlEx input file
}
