% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{exportGenoPop}
\alias{exportGenoPop}
\title{write a genepop input file}
\usage{
exportGenoPop(
  x,
  filename,
  header = "genePop file",
  pops = NULL,
  loci = NULL,
  useIndNames = FALSE
)
}
\arguments{
\item{x}{an EFGLdata object}

\item{filename}{the name of the file to write}

\item{header}{a string to use as the header line of the genepop file}

\item{pops}{a vector of pops to include. If not specified,
all pops are used.}

\item{loci}{a vector of loci to include. If not specified,
all loci are used.}

\item{useIndNames}{TRUE to use individual names as sample identifiers. Otherwise,
population names are used}

\item{return}{nothing, just writes a file}
}
\description{
write a genepop input file
}
