% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{exportHierFstat}
\alias{exportHierFstat}
\title{export a hierfstat input dataframe}
\usage{
exportHierFstat(x, pops = NULL, loci = NULL)
}
\arguments{
\item{x}{an EFGLdata object}

\item{pops}{a vector of pops to include. If not specified, all pops are used.}

\item{loci}{a vector of loci to use. If not specified,
all loci are used.}
}
\value{
a dataframe coded to be used as input for hierfstat
}
\description{
export a hierfstat input dataframe
}
