% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{exportPlink}
\alias{exportPlink}
\title{export a Plink PED file and optionally a (not very useful except as a template) MAP file}
\usage{
exportPlink(
  x,
  filename,
  pops = NULL,
  loci = NULL,
  FID = "Ind",
  IID = "Ind",
  pa = NULL,
  ma = NULL,
  sex = NULL,
  pheno = NULL,
  map = NULL
)
}
\arguments{
\item{x}{an EFGLdata object}

\item{filename}{the name of the file to write}

\item{pops}{a vector of pops to include. If not specified, all pops are used.}

\item{loci}{a vector of loci to use. If not specified,
all loci are used.}

\item{FID}{Metadata column name to use as the family ID}

\item{IID}{Metadata column name to use as the within-family ID}

\item{pa}{Metadata column name to use as the within-family ID of the father.
If NULL, all are listed as unknown (0).}

\item{ma}{Metadata column name to use as the within-family ID of the mother.
If NULL, all are listed as unknown (0).}

\item{sex}{Metadata column to use as the sex. This should be coded as "M"
for male, "F" for female, and any other values are treated as Unknown.
If NULL, all are listed as unknown (0).}

\item{pheno}{Metadata column to use as the phenotype. This should be coded
as "1" for control, "2" for case, and "-9" or "0" for Unknown. If NULL,
all are listed as unknown (0).}

\item{map}{filename to write a MAP file. If NULL, no MAP file is written.
This just writes a dummy MAP file with the loci names in order and all
given the same chromosome code and positions of "0".
If you need a valid MAP file, you will need to edit this.}
}
\value{
nothing, just writes a file
}
\description{
Only biallelic markers are used, exports as two column per call. This format can be used
as the input for "Admixture".
}
