% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{exportProgenyStyle}
\alias{exportProgenyStyle}
\title{export a "Progeny-style" export file for later reading into EFGLmh}
\usage{
exportProgenyStyle(x, filename, pops = NULL, loci = NULL, metadata = NULL)
}
\arguments{
\item{x}{an EFGLdata object}

\item{filename}{the name of the file to write}

\item{pops}{a vector of pops to include. If not specified, all pops are used.}

\item{loci}{a vector of loci to use. If not specified,
all loci are used.}

\item{metadata}{a vector of metadata fields to include. If not specified,
all fields are used.}
}
\value{
nothing, just writes a file
}
\description{
Columns in order are Pop, Ind, metadata, genotypes (2-column per call)
Missing genotypes are "0" for SNPs (biallelic or nonvariable with alleles represented by 1 character)
and "000" for others. If a locus is all missing, it is treated as a SNP. Pop column is called
"Pedigree" and Ind column is called "Individual.Name".
}
