% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{exportRubias_baseline}
\alias{exportRubias_baseline}
\title{export a rubias baseline}
\usage{
exportRubias_baseline(
  x,
  pops = NULL,
  repunit = NULL,
  collection = NULL,
  loci = NULL
)
}
\arguments{
\item{x}{an EFGLdata object}

\item{pops}{a vector of pops to include in the baseline. If not specified, all pops are used.}

\item{repunit}{the column name of the metadata variable designating repunit.
This can be Pop to use the population name. If not specified, NA is used for all samples.}

\item{collection}{the column name of the metadata variable designating collection.
This can be Pop to use the population name. If not specified, NA is used for all samples.}

\item{loci}{a vector of loci to use. If not specified,
all loci are used.}
}
\value{
a tibble
}
\description{
export a rubias baseline
}
