% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{exportRubias_mixture}
\alias{exportRubias_mixture}
\title{export a rubias mixture}
\usage{
exportRubias_mixture(x, pops = NULL, collection = NULL, loci = NULL)
}
\arguments{
\item{x}{an EFGLdata object}

\item{pops}{a vector of pops to include in the baseline. If not specified, all pops are used.}

\item{collection}{the column name of the metadata variable designating collection.
This can be Pop to use the population name. If not specified, NA is used for all samples.
For mixtures, this variable indicates what samples come from the same "stratum" - to be analyzed together.}

\item{loci}{a vector of loci to use. If not specified,
all loci are used.}
}
\value{
a tibble
}
\description{
export a rubias mixture
}
