% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{exportSNPPIT}
\alias{exportSNPPIT}
\title{write a SNPPIT input file. Will warn about skipping loci with > 2 alleles.}
\usage{
exportSNPPIT(
  x,
  filename,
  baseline,
  mixture,
  loci = NULL,
  errorRate = 0.005,
  POPCOLUMN_SEX = NULL,
  POPCOLUMN_REPRO_YEARS = NULL,
  POPCOLUMN_SPAWN_GROUP = NULL,
  OFFSPRINGCOLUMN_BORN_YEAR = NULL,
  OFFSRPINGCOLUMN_SAMPLE_YEAR = NULL,
  OFFSPRINGCOLUMN_AGE_AT_SAMPLING = NULL
)
}
\arguments{
\item{x}{an EFGLdata object}

\item{filename}{the name of the file to write}

\item{baseline}{a vector of pops to use as the baseline (potential parents).}

\item{mixture}{a vector of pops to use as the mixture (potential offspring).}

\item{loci}{a vector of loci to include. If not specified,
all loci are used.}

\item{errorRate}{per allele error rate for all loci}

\item{POPCOLUMN_SEX}{metadata column with sex info (coded as M, F, and ?)}

\item{POPCOLUMN_REPRO_YEARS}{metadata column with repro years}

\item{POPCOLUMN_SPAWN_GROUP}{metadata column with spawn group}

\item{OFFSPRINGCOLUMN_BORN_YEAR}{metadata column with birth year}

\item{OFFSRPINGCOLUMN_SAMPLE_YEAR}{metadata column with sample year}

\item{OFFSPRINGCOLUMN_AGE_AT_SAMPLING}{metadata column with age at sampling}
}
\value{
nothing, just writes a file
}
\description{
write a SNPPIT input file. Will warn about skipping loci with > 2 alleles.
}
