% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{exportStructure}
\alias{exportStructure}
\title{write a Structure input file.}
\usage{
exportStructure(x, filename, pops = NULL, loci = NULL)
}
\arguments{
\item{x}{an EFGLdata object}

\item{filename}{the name of the file to write}

\item{pops}{a vector of pops to include. If not specified,
all pops are used.}

\item{loci}{a vector of loci to include. If not specified,
all loci are used.}
}
\value{
nothing, just writes a file
}
\description{
Columns are Ind, Pop (converted to integers), then genotypes. Missing alleles
are coded as -9
}
