% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_EFGLdata.R
\name{getInds}
\alias{getInds}
\title{get a vector of individuals present}
\usage{
getInds(x, pops = NULL)
}
\arguments{
\item{x}{an EFGLdata object}

\item{pops}{a vector of pops that you want individual names for. If not
specified, names for all pops are returned}
}
\value{
a vector of the Individual names present
}
\description{
get a vector of individuals present
}
