% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_EFGLdata.R
\name{getMeta}
\alias{getMeta}
\title{get a vector of metadata column names present}
\usage{
getMeta(x)
}
\arguments{
\item{x}{an EFGLdata object}
}
\description{
get a vector of metadata column names present
}
