% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_EFGLdata.R
\name{getPops}
\alias{getPops}
\title{get a vector of populations (pedigrees) present}
\usage{
getPops(x)
}
\arguments{
\item{x}{an EFGLdata object}
}
\value{
a vector of the unique population names present
}
\description{
get a vector of populations (pedigrees) present
}
