% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_EFGLdata.R
\name{lociSuccess}
\alias{lociSuccess}
\title{calculate genotyping success of loci (uses only allele 1 for each
  genotype - assumes if allele 1 is (is not) NA, so is (is not) allele 2)}
\usage{
lociSuccess(x)
}
\arguments{
\item{x}{an EFGLdata object}
}
\value{
a tibble giving the genotyping success of each locus as a proportion
}
\description{
calculate genotyping success of loci (uses only allele 1 for each
  genotype - assumes if allele 1 is (is not) NA, so is (is not) allele 2)
}
