% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_EFGLdata.R
\name{moveInds}
\alias{moveInds}
\title{combine individuals into one population AND REMOVE the previous entry
  for those individuals}
\usage{
moveInds(x, inds, newName)
}
\arguments{
\item{x}{an EFGLdata object}

\item{inds}{a vector of individuals to put in the new pop}

\item{newName}{a string giving the name of population to add the individuals
too. This can be a new pop or an existing pop (a warning is issued if existing).}
}
\value{
an EFGLdata object
}
\description{
combine individuals into one population AND REMOVE the previous entry
  for those individuals
}
