% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_EFGLdata.R
\name{movePops}
\alias{movePops}
\title{combine populations into one AND REMOVE the old populations}
\usage{
movePops(x, pops, newName)
}
\arguments{
\item{x}{an EFGLdata object}

\item{pops}{a vector of populations to combine}

\item{newName}{a string giving the name of the population to combine pops
into. This can be a new pop or an existing pop (a warning is issued if existing).}
}
\value{
an EFGLdata object
}
\description{
combine populations into one AND REMOVE the old populations
}
