% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFGLdata.R
\name{mtype2wide}
\alias{mtype2wide}
\title{A function for turning the genotype output of mtype2 (microTyper)
into a wide format for input into EFGLmh}
\usage{
mtype2wide(x, popName = "PopulationName")
}
\arguments{
\item{x}{the output of mtype2, either a tibble or a path to a file}

\item{popName}{The value to give all samples for population name.}
}
\description{
A function for turning the genotype output of mtype2 (microTyper)
into a wide format for input into EFGLmh
}
