% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_EFGLdata.R
\name{numInds}
\alias{numInds}
\title{get the number of individuals present in each pop}
\usage{
numInds(x, pops = NULL)
}
\arguments{
\item{x}{an EFGLdata object}

\item{pops}{a vector of pops that you want individual names for. If not
specified, names for all pops are returned}
}
\value{
a named vector with the number of individuals in each pop
}
\description{
get the number of individuals present in each pop
}
