% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_EFGLdata.R
\name{removeInds}
\alias{removeInds}
\title{remove individuals from an EFGLdata object}
\usage{
removeInds(x, inds)
}
\arguments{
\item{x}{an EFGLdata object}

\item{inds}{a vector of individuals to remove}
}
\value{
an EFGLdata object
}
\description{
remove individuals from an EFGLdata object
}
