% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_EFGLdata.R
\name{removeLoci}
\alias{removeLoci}
\title{remove loci from an EFGLdata object}
\usage{
removeLoci(x, lociRemove)
}
\arguments{
\item{x}{an EFGLdata object}

\item{lociRemove}{a vector of loci names to remove}
}
\description{
remove loci from an EFGLdata object
}
