% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_EFGLdata.R
\name{removePops}
\alias{removePops}
\title{remove pops from an EFGLdata object}
\usage{
removePops(x, pops)
}
\arguments{
\item{x}{an EFGLdata object}

\item{pops}{a vector of pops to remove}
}
\value{
an EFGLdata object
}
\description{
remove pops from an EFGLdata object
}
