% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_EFGLdata.R
\name{whichLower}
\alias{whichLower}
\title{Identify which individual out of duplicate pairs has the lower genotyping success}
\usage{
whichLower(dupTable, geno_success)
}
\arguments{
\item{dupTable}{the output of close_matching_samples (from rubias)}

\item{geno_success}{the output of genoSuccess}
}
\value{
a vector of unique individual names representing the individuals with lower
  genotyping success from each pair
}
\description{
Identify which individual out of duplicate pairs has the lower genotyping success
}
