% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_EFGLdata.R
\name{calcAF}
\alias{calcAF}
\title{calculate allele frequencies within populations
Combinations of pop, locus, allele that are not observed are
(silently) excluded from the output. To add these back in, one can
use the `tidyverse` function `complete` with the output of this function.}
\usage{
calcAF(x)
}
\arguments{
\item{x}{an EFGLdata object}
}
\value{
a tibble
}
\description{
calculate allele frequencies within populations
Combinations of pop, locus, allele that are not observed are
(silently) excluded from the output. To add these back in, one can
use the `tidyverse` function `complete` with the output of this function.
}
