% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{exportGenAlEx}
\alias{exportGenAlEx}
\title{write a GenAlEx input file. But you really should not be using GenAlEx.}
\usage{
exportGenAlEx(x, filename, pops = NULL, loci = NULL, title = "")
}
\arguments{
\item{x}{an EFGLdata object}

\item{filename}{the name of the file to write}

\item{pops}{a vector of pops to include. If not specified,
all pops are used.}

\item{loci}{a vector of loci to include. If not specified,
all loci are used.}

\item{title}{a string to use as the "title" row}

\item{return}{nothing, just writes a file}
}
\description{
write a GenAlEx input file. But you really should not be using GenAlEx.
}
